правки в data.sql

+++ src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 -- Таблица пользователей
 CREATE TABLE IF NOT EXISTS users (
-    user_id SERIAL PRIMARY KEY,
+    user_id BIGINT PRIMARY KEY,
     email VARCHAR(255) NOT NULL UNIQUE,
     login VARCHAR(50) NOT NULL UNIQUE,
     name VARCHAR(100) NOT NULL,
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS users (
 
 -- Таблица фильмов
 CREATE TABLE IF NOT EXISTS films (
-    film_id SERIAL PRIMARY KEY,
+    film_id BIGINT PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT,
     release_date DATE NOT NULL,
@@ -19,14 +19,14 @@ CREATE TABLE IF NOT EXISTS films (
 
 -- Таблица жанров
 CREATE TABLE IF NOT EXISTS genres (
-    genre_id SERIAL PRIMARY KEY,
+    genre_id BIGINT PRIMARY KEY,
     name VARCHAR(50) NOT NULL UNIQUE
 );
 
 -- Таблица дружбы
 CREATE TABLE IF NOT EXISTS friendship (
-    user_id INTEGER NOT NULL,
-    friend_id INTEGER NOT NULL,
+    user_id BIGINT NOT NULL,
+    friend_id BIGINT NOT NULL,
     status VARCHAR(20) NOT NULL,
     PRIMARY KEY (user_id, friend_id),
     FOREIGN KEY (user_id) REFERENCES users(user_id),
@@ -35,8 +35,8 @@ CREATE TABLE IF NOT EXISTS friendship (
 
 -- Таблица лайков
 CREATE TABLE IF NOT EXISTS likes (
-    user_id INTEGER NOT NULL,
-    film_id INTEGER NOT NULL,
+    user_id BIGINT NOT NULL,
+    film_id BIGINT NOT NULL,
     PRIMARY KEY (user_id, film_id),
     FOREIGN KEY (user_id) REFERENCES users(user_id),
     FOREIGN KEY (film_id) REFERENCES films(film_id)
@@ -44,8 +44,8 @@ CREATE TABLE IF NOT EXISTS likes (
 
 -- Таблица жанров фильмов
 CREATE TABLE IF NOT EXISTS film_genres (
-    film_id INTEGER NOT NULL,
-    genre_id INTEGER NOT NULL,
+    film_id BIGINT NOT NULL,
+    genre_id BIGINT NOT NULL,
     PRIMARY KEY (film_id, genre_id),
     FOREIGN KEY (film_id) REFERENCES films(film_id),
     FOREIGN KEY (genre_id) REFERENCES genres(genre_id)
